requirement.txt
fastapi
uvicorn
langchain
langchain-community
langgraph
langchain-ollama
pydantic
paho-mqtt
influxdb-client-3
python-dotenv
requests